requirements.txt
matplotlib==3.6.2
nltk==3.7
pandas==1.2.0
Pillow==9.4.0
plotly==4.14.3
streamlit==1.9.0
wordcloud==1.8.2.2
